User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -65,6 +65,11 @@ class ListingController extends Controller
     // Update Listing Data
     public function update(Request $request, Listing $listing)
     {
+        // Make Sure Logged In User Is Owner
+        if($listing->user_id != auth()->id()) {
+           abort('403', 'Unauthorized Action'); 
+        }
+
         $formFields = $request->validate([
              'title' => 'required',
              'company' => 'required',
@@ -87,6 +92,11 @@ class ListingController extends Controller
     // Delete Listing 
     public function destroy(Listing $listing)
     {
+         // Make Sure Logged In User Is Owner
+         if($listing->user_id != auth()->id()) {
+            abort('403', 'Unauthorized Action'); 
+         }
+         
         $listing->delete();
         
         return redirect('/')->with('message', 'Listing Deleted!');
